refactor(runtime): separate finish assembly and boot manifest tests
Keep the pipeline and root test modules below the file-size guard after cleanup receipt support, without changing settlement behavior.
